Fix use of const in declarations
This is a big commit, needed to get ridd of the warnings about const
declarations. A log of variables and data structures used const, but was
used in calls to functions that modified the data. Theese warnings can be
real errors, and should be avoided. On the other hand, the compiler will
usualy optimize quite well without const declarations. None of the changes
should modifiy the code's behaviour.

Jan KÃre Vatne